Chore: 배너 컴포넌트 분리
